Moved book list login into children
